Splitted Simulation methods in multiple files
