+++ AeViz/simulation/methods/magnetic_fields.py
@@ -1,5 +1,4 @@
-import numpy as np
-from AeViz.utils.decorators import hdf_isopen, derive, smooth
+from AeViz.simulation.methods import *
 from AeViz.utils.math_utils import strfct2D
 
 ## -----------------------------------------------------------------
