+++ AeViz/simulation/methods/thermo.py
@@ -1,5 +1,4 @@
-import numpy as np
-from AeViz.utils.decorators import hdf_isopen, derive, smooth
+from AeViz.simulation.methods import *
 
 ## -----------------------------------------------------------------
 ## THERMODYNAMICAL DATA
